TESTE_OPA_GAMES/requirements.txt
asyncio
websockets
mysql-connector-python
keyboard